Refactor audio engine and physics systems
- Removed unnecessary forward declarations in audio_engine.cpp and physics_world.cpp.
- Improved type safety in audio_engine.cpp by ensuring sample rate is cast to float before calculations.
- Enhanced event dispatcher to allow raw function pointer subscriptions with handler IDs.
- Cleaned up unused vector normalization and cross product functions in boat.cpp.
- Updated buoyancy calculations to ignore body rotation where not needed and improved voxel buoyancy calculations for clarity.
- Simplified character controller system function signature by removing unused parameters.
- Optimized transform handling in scene system by introducing a PreviousTransform struct to store TRS data.
- Improved entity pooling logic in entity_pool.cpp for better performance and clarity.
- Enhanced hierarchy management in hierarchy.cpp to cache children for efficiency.
- Updated scene serialization to ensure proper type casting for asset handles.
- Added tests for PreviousTransform to validate new functionality.

+++ engine/audio/src/miniaudio_impl.cpp
@@ -822,7 +822,7 @@ void set_music_position_impl(AudioEngine::Impl* impl, MusicHandle h, float secon
     if (it == impl->music.end() || !it->second.loaded) return;
 
     ma_uint32 sample_rate = ma_engine_get_sample_rate(&impl->engine);
-    ma_uint64 frame = static_cast<ma_uint64>(seconds * sample_rate);
+    ma_uint64 frame = static_cast<ma_uint64>(seconds * static_cast<float>(sample_rate));
     ma_sound_seek_to_pcm_frame(&it->second.sound, frame);
 }
